requirements/requirements.txt
brax>=0.9.0
chex
colorama
craftax
distrax @ git+https://github.com/google-deepmind/distrax  # distrax release doesn't support jax > 0.4.13
envpool; sys_platform == "linux"
flashbax @ git+https://github.com/instadeepai/flashbax
flax
gymnasium
gymnax>=0.0.6
huggingface_hub
hydra-core==1.3.2
id-marl-eval @ git+https://github.com/EdanToledo/marl-eval
# jax>=0.4.25
# jaxlib
jaxmarl
jumanji==1.0.0
mctx
navix
neptune
numpy
omegaconf
optax @ git+https://github.com/google-deepmind/optax.git
pgx
popjym @ git+https://github.com/FLAIROx/popjym.git
protobuf==3.20.2
rlax
tensorboard_logger
tensorflow_probability
tqdm
wandb
xminigrid @ git+https://github.com/corl-team/xland-minigrid.git@main
